move the `LocationTypeDataFileTypeValidators` and the `OperandInterpolationValidators` to the validators subpackage
