+++ hydrolib/core/dflowfm/validators/__init__.py
@@ -1,5 +1,13 @@
 """Shared validator mixins for D-Flow FM block models."""
 
-from hydrolib.core.dflowfm.validators.fields import CoordinateValidator
+from hydrolib.core.dflowfm.validators.fields import (
+    CoordinateValidator,
+    LocationTypeDataFileTypeValidators,
+    OperandInterpolationValidators,
+)
 
-__all__ = ["CoordinateValidator"]
+__all__ = [
+    "CoordinateValidator",
+    "OperandInterpolationValidators",
+    "LocationTypeDataFileTypeValidators",
+]
